Refactor nodemailer/mailgun
