+++ .gitignore
@@ -3,4 +3,4 @@ public/build
 .sass-cache
 assets/video/
 node_modules
-config.js
+server/config.js
