+++ server.js
@@ -1,7 +1,7 @@
 var express = require('express');
 var bodyParser = require('body-parser');
 var cors = require('cors');
-var router = require('./router');
+var router = require('./server/router');
 
 var app = express();
 var port = 3000;
