+++ server/mail.js
@@ -1,0 +1,17 @@
+var nodemailer = require('nodemailer');
+var mailgun = require('nodemailer-mailgun-transport');
+
+if (!process.env.MY_EMAIL) {
+  var config = require('./config');
+}
+
+var auth = {
+  auth: {
+    api_key: process.env.MAILGUN_API || config.MAILGUN_API,
+    domain: process.env.MAILGUN_DOMAIN || config.MAILGUN_DOMAIN
+  }
+};
+
+var transporter = nodemailer.createTransport(mailgun(auth));
+
+module.exports = transporter;
