+++ server/router.js
@@ -1,20 +1,6 @@
 var router = require('express').Router();
 var path = require('path');
-var nodemailer = require('nodemailer');
-var mailgun = require('nodemailer-mailgun-transport');
-
-if (!process.env.MY_EMAIL) {
-  var config = require('./config');
-}
-
-var auth = {
-  auth: {
-    api_key: process.env.MAILGUN_API || config.MAILGUN_API,
-    domain: process.env.MAILGUN_DOMAIN || config.MAILGUN_DOMAIN
-  }
-};
-
-var transporter = nodemailer.createTransport(mailgun(auth));
+var mail = require('./mail');
 
 router.get('/*', function(req, res) {
   res.sendFile(path.join(__dirname, '/public/', req.url));
@@ -28,7 +14,7 @@ router.post('/contact', function(req, res) {
     text: req.body.text
   };
 
-  transporter.sendMail(message, function(err, info) {
+  mail.sendMail(message, function(err, info) {
     if(err) {
       console.log('error');
       res.json({success: false});
